내용 : 2단계 - 로또(수동) - 코드리뷰 요청(테스트 코드 수정) 이슈 : -

+++ src/main/java/lotto/domain/EqualNumbersMoney.java
@@ -32,6 +32,6 @@ public enum EqualNumbersMoney {
     }
 
     public String toString() {
-        return this.equalNumbers + "개 일치 (" + this.winOfMoney + "원) - " + this.equalCount + "개";
+        return this.equalNumbers + "개 일치 (" + (int)this.winOfMoney + "원) - " + this.equalCount + "개";
     }
 }
